% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{adjust}
\alias{adjust}
\title{Compute adjusted dates}
\usage{
adjust(dates, bdc = 0L)
}
\arguments{
\item{dates}{A Date vector with dates}

\item{bdc}{An integer corresponding to the business-day convetion}
}
\value{
A Date vector with dates adjust according to business-day convention
}
\description{
Adjust a vector of dates following a business-day convention
}
\details{
This function takes a vector of dates and returns another vector of dates
of the same length returning at each position the adjusted date according
to the selected business-day convention matching the existing QuantLib
enumeration type. Currently supported values are (starting from zero):
\sQuote{Following}, \sQuote{ModifiedFollowing}, \sQuote{Preceding},
\sQuote{ModifiedPreceding}, \sQuote{Unadjusted}, \sQuote{HalfModifiedFollowing}
\sQuote{Nearest} and a fallback of \sQuote{Unadjusted} for all other value.
}
\examples{
getEndOfMonth(Sys.Date()+0:6)
}
