% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{advanceDate}
\alias{advanceDate}
\title{Advance a date}
\usage{
advanceDate(rd, days = 0L)
}
\arguments{
\item{rd}{A Date object describing the date to be advanced to the
next business day.}

\item{days}{An optional integer offset applied to the date}
}
\value{
The advanced date is returned
}
\description{
Advance a date to the next business day plus an optional shift
}
\details{
This function takes a given date and advances it to the next business day
under the current (global) calendar setting. If an optional offset value is
given it is applied as well.
}
\examples{
advanceDate(Sys.Date(), 2)  # today to the next biz day, plus 2 days
}
