% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getHolidays}
\alias{getHolidays}
\title{Compute holidays}
\usage{
getHolidays(from, to, includeWeekends = FALSE)
}
\arguments{
\item{from}{A Date object with the start date}

\item{to}{A Date object with the end date}

\item{includeWeekends}{A boolean indicating if weekends should be included, default
is \sQuote{FALSE}}
}
\value{
A Date vector with holidays between the given dates
}
\description{
Compute the number of holidays between two dates
}
\details{
This function takes a start and end date and returns a vector of holidays
between them according to the active calendar.
}
\examples{
getHolidays(Sys.Date(), Sys.Date() + 30)
}
