% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getName}
\alias{getName}
\alias{getId}
\title{Get calendar name, or id}
\usage{
getName()

getId()
}
\value{
A string with the calendar name
}
\description{
Get calendar name or id
}
\details{
This function returns the corresponding (full) name (as in the underlying
implementationclass) or identification string (used to select it) of the
current calendar.
}
\examples{
getName()
}
