% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{isBusinessDay}
\alias{isBusinessDay}
\title{Test for business days}
\usage{
isBusinessDay(dates)
}
\arguments{
\item{dates}{A Date vector with dates to be examined}
}
\value{
A logical vector indicating which dates are business days
}
\description{
Test a vector of dates for business day
}
\details{
This function takes a vector of dates and returns a logical vector
of the same length indicating at each position whether the corresponding
date is a business day in the currently active (global) calendar.
}
\examples{
isBusinessDay(Sys.Date()+0:6)
}
