% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{isEndOfMonth}
\alias{isEndOfMonth}
\title{Test for end-of-month}
\usage{
isEndOfMonth(dates)
}
\arguments{
\item{dates}{A Date vector with dates to be examined}
}
\value{
A logical vector indicating which dates are end-of-month
}
\description{
Test a vector of dates for end-of-month
}
\details{
This function takes a vector of dates and returns a logical vector
of the same length indicating at each position whether the corresponding
date is at the end of a month in the currently active (global) calendar.
}
\examples{
isEndOfMonth(Sys.Date()+0:6)
}
