% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{isWeekend}
\alias{isWeekend}
\title{Test for weekends}
\usage{
isWeekend(dates)
}
\arguments{
\item{dates}{A Date vector with dates to be examined}
}
\value{
A logical vector indicating which dates are weekends
}
\description{
Test a vector of dates for weekends
}
\details{
This function takes a vector of dates and returns a logical vector
of the same length indicating at each position whether the corresponding
date is a weekend in the currently active (global) calendar.
}
\examples{
isWeekend(Sys.Date()+0:6)
}
