\name{RcppQuantuccia-package}
\alias{RcppQuantuccia-package}
\alias{RcppQuantuccia}
\docType{package}
\title{\packageTitle{RcppQuantuccia}}
\description{\packageDescription{RcppQuantuccia}}
\details{
  The DESCRIPTION file: \packageDESCRIPTION{RcppQuantuccia}
  \packageIndices{RcppQuantuccia}

  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
  \packageAuthor{RcppQuantuccia}
  
  Maintainer: \packageMaintainer{RcppQuantuccia}
}
\references{
  This optional section can contain literature or other references for
  background information.
}
% Optionally other standard keywords, one per line,
% from the file KEYWORDS in the R documentation.
\keyword{package}
\seealso{
  Optional links to other man pages
}
\examples{
  ## Optional simple examples of the most important functions
  ## Use \dontrun{} around code to be shown but not executed
}
