% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{setCalendar}
\alias{setCalendar}
\title{Set a calendar}
\usage{
setCalendar(calstr)
}
\arguments{
\item{calstr}{A character variable containing the market for which a calendar
is to be set}
}
\value{
Nothing is returned but the global state is changed
}
\description{
Set a calendar
}
\details{
This function sets a calendar to the given market or country convention.
Note that at present only the default \sQuote{TARGET} and \sQuote{UnitedStates}
are supported.
}
\examples{
setCalendar("UnitedStates")
}
