\name{RcppQuantuccia-package}
\alias{RcppQuantuccia-package}
\alias{RcppQuantuccia}
\docType{package}
\title{\packageTitle{RcppQuantuccia}}
\description{\packageDescription{RcppQuantuccia}}
\details{The DESCRIPTION file: \packageDESCRIPTION{RcppQuantuccia}}
\section{Package Content}{\packageIndices{RcppQuantuccia}}
\author{\packageAuthor{RcppQuantuccia}}
\section{Maintainer}{\packageMaintainer{RcppQuantuccia}}
\references{\url{https://www.quantlib.org/}}
\keyword{package}
